# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 단짝 공동 HTTP 계약
  version: 1.0.0
  description: '#11에서 확정한 구현 기준입니다. API 실행 완료를 뜻하지 않습니다. 업무 규칙은 docs/specs/requirements, 저장·종료 책임은 contracts/decisions.md를
    따릅니다. 모든 식별자는 십진 문자열 또는 명시된 UUID, 금액은 정수 원입니다. 오류는 ApiError를 사용합니다. 개인정보 응답은 Cache-Control: no-store. 인증 쿠키는
    HttpOnly이며 운영 HTTPS에서는 Secure, SameSite=Lax. 상태 변경 요청은 세션의 X-CSRF-Token을 검증합니다.'
  contact:
    name: 단짝 개발팀
    url: https://github.com/team-bestfriend/danjjak-app/issues/11
servers:
- url: /
  description: 같은 origin 또는 FE 개발 프록시. Tomcat context가 /danjjak이면 FE base URL에 /danjjak을 지정.
security:
- SessionCookie: []
...
paths:
...
  /api/auth/session:
    get:
      operationId: getSession
      tags:
      - auth
      summary: 세션과 CSRF 값 확인
      description: 미로그인은 200 authenticated=false,user=null. 보호 업무 API의 미인증은 401. 세션 확인 자체의 서버 실패와 미로그인을 구분합니다.
      x-requirements:
      - FR-057
      responses:
        '200':
          description: 처리 성공.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              examples:
                sample:
                  $ref: ./examples/session-anonymous.json
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '400':
          $ref: '#/components/responses/Error400'
        '500':
          $ref: '#/components/responses/Error500'
      security: []
...
  /api/support:
    get:
      operationId: getSupport
      tags:
      - support
      summary: 고객센터·보호자 연락처 조회
      description: 동의 선택 미완료는 409. 선택 동의는 모두 거절할 수 있습니다. 고객센터 제공 번호가 설정되지 않으면 503. 보호자 미등록은 guardian.phoneNumber=null. 전화번호로 카카오 수신자를 정하지 않습니다.
      x-requirements:
      - FR-009
      - FR-034
      - FR-043
      responses:
        '200':
          description: 처리 성공.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Support'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '409':
          $ref: '#/components/responses/Error409'
        '500':
          $ref: '#/components/responses/Error500'
        '503':
          $ref: '#/components/responses/Error503'
  /api/support/guardian:
    put:
      operationId: updateGuardian
      tags:
      - support
      summary: 보호자 전화번호 저장
      description: 보호자 전화번호 저장
      x-requirements:
      - FR-009
      responses:
        '200':
          description: 처리 성공.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guardian'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '500':
          $ref: '#/components/responses/Error500'
      security:
      - SessionCookie: []
        CsrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuardianUpdate'
  /api/anomaly-events/{anomalyEventId}/guardian-notification:
    post:
      operationId: notifyGuardianDemo
      tags:
      - support
      summary: 카카오 본인 알림 시연
      description: 본인·미결정·HIGH·guardianSharing 동의·명시적 선택 모두 검증. 동의 없으면403이며 실제/모의 전송 모두 금지. 실제 성공만 actualNotifiedAt
        저장, 같은 미결정 판정은 성공 결과 재사용. 자격 없음/명확한 실패 후 모의는 sentAt=null, 영속 전송 이력 없음. 외부 시간 초과로 결과 불명은502 NOTIFICATION_RESULT_UNKNOWN이며
        자동 재전송하지 않습니다. 결정 후/낮은 위험은409. 알림 실패가 금융 결과를 변경하지 않습니다.
      x-requirements:
      - FR-042
      - FR-053
      responses:
        '200':
          description: 처리 성공.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationResult'
          headers:
            Cache-Control:
              schema:
                type: string
                enum:
                - no-store
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '404':
          $ref: '#/components/responses/Error404'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '500':
          $ref: '#/components/responses/Error500'
        '502':
          $ref: '#/components/responses/Error502'
      security:
      - SessionCookie: []
        CsrfToken: []
      parameters:
      - name: anomalyEventId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotificationRequest'
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: 카카오 성공 후 서버 세션. OAuth 토큰은 FE에 반환하지 않습니다.
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: GET /api/auth/session에서 받은 세션 CSRF 값. 로그인/로그아웃 시 재발급.
  schemas:
    Id:
      type: string
      pattern: ^[1-9][0-9]{0,18}$
      description: DB BIGINT 식별자의 십진 문자열. 숫자로 변환하지 않습니다.
      example: '101'
...
    Instant:
      type: string
      format: date-time
      description: UTC RFC 3339 시각. 저장은 UTC, 날짜 집계는 Asia/Seoul.
...
    VoiceMode:
      type: string
      enum:
      - TTS
      - FAMILY
    SpeechSpeed:
      type: string
      enum:
      - SLOW
      - NORMAL
      - FAST
...
    PhoneNumber:
      type: string
      pattern: ^[0-9]+(?:-[0-9]+)*$
      minLength: 5
      maxLength: 30
      description: 숫자 또는 숫자 묶음 사이 하이픈. 카카오 수신자 식별자가 아닙니다.
    ApiError:
      type: object
      additionalProperties: false
      required:
      - code
      - message
      - fieldErrors
      - retryable
      - requestId
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          minLength: 1
          description: 사용자에게 보여 줄 쉬운 한국어. 내부 예외·입력값을 포함하지 않습니다.
        fieldErrors:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
            - field
            - code
            - message
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string
        retryable:
          type: boolean
        requestId:
          type: string
          format: uuid
      example:
        code: PIN_MISMATCH
        message: 비밀번호가 맞지 않아요. 다시 입력해 주세요.
        fieldErrors: []
        retryable: true
        requestId: 00000000-0000-4000-8000-000000000001
...
    Consents:
      type: object
      additionalProperties: false
      required:
      - completed
      - usageRecording
      - guardianSharing
      properties:
        completed:
          type: boolean
        usageRecording:
          type: boolean
        guardianSharing:
          type: boolean
...
    Settings:
      type: object
      additionalProperties: false
      required:
      - fontSize
      - speechSpeed
      - defaultVoiceMode
      properties:
        fontSize:
          type: string
          enum:
          - SMALL
          - NORMAL
          - LARGE
        speechSpeed:
          $ref: '#/components/schemas/SpeechSpeed'
        defaultVoiceMode:
          $ref: '#/components/schemas/VoiceMode'
    CurrentUser:
      type: object
      additionalProperties: false
      required:
      - userId
      - name
      - consents
      - settings
      - accountReady
      properties:
        userId:
          $ref: '#/components/schemas/Id'
        name:
          type: string
          minLength: 1
          maxLength: 50
          pattern: \S
        consents:
          $ref: '#/components/schemas/Consents'
        settings:
          $ref: '#/components/schemas/Settings'
        accountReady:
          type: boolean
    AnonymousSession:
      type: object
      additionalProperties: false
      required:
      - authenticated
      - user
      - csrfToken
      properties:
        authenticated:
          type: boolean
          enum:
          - false
        user:
          type: object
          nullable: true
          enum:
          - null
        csrfToken:
          type: string
          minLength: 1
    AuthenticatedSession:
      type: object
      additionalProperties: false
      required:
      - authenticated
      - user
      - csrfToken
      properties:
        authenticated:
          type: boolean
          enum:
          - true
        user:
          $ref: '#/components/schemas/CurrentUser'
        csrfToken:
          type: string
          minLength: 1
    Session:
      oneOf:
      - $ref: '#/components/schemas/AnonymousSession'
      - $ref: '#/components/schemas/AuthenticatedSession'
...
    NotificationRequest:
      type: object
      additionalProperties: false
      required:
      - confirmedSelfDemo
      properties:
        confirmedSelfDemo:
          type: boolean
          enum:
          - true
      description: 화면에서 본인 계정 나에게 보내기 시연을 확인한 뒤 제출.
    ActualNotification:
      type: object
      additionalProperties: false
      required:
      - mode
      - message
      - sentAt
      - recipient
      properties:
        mode:
          type: string
          enum:
          - ACTUAL
        message:
          type: string
        sentAt:
          $ref: '#/components/schemas/Instant'
        recipient:
          type: string
          enum:
          - SELF
    MockNotification:
      type: object
      additionalProperties: false
      required:
      - mode
      - message
      - sentAt
      - recipient
      properties:
        mode:
          type: string
          enum:
          - MOCK_NO_CREDENTIALS
          - MOCK_AFTER_FAILURE
        message:
          type: string
        sentAt:
          type: string
          nullable: true
          enum:
          - null
        recipient:
          type: string
          enum:
          - SELF
    NotificationResult:
      oneOf:
      - $ref: '#/components/schemas/ActualNotification'
      - $ref: '#/components/schemas/MockNotification'
    Guardian:
      type: object
      additionalProperties: false
      required:
      - phoneNumber
      properties:
        phoneNumber:
          type: string
          pattern: ^[0-9]+(?:-[0-9]+)*$
          minLength: 5
          maxLength: 30
          description: 숫자 또는 숫자 묶음 사이 하이픈. 카카오 수신자 식별자가 아닙니다.
          nullable: true
    GuardianUpdate:
      type: object
      additionalProperties: false
      required:
      - phoneNumber
      properties:
        phoneNumber:
          $ref: '#/components/schemas/PhoneNumber'
    Support:
      type: object
      additionalProperties: false
      required:
      - guardian
      - customerCenter
      - notificationRecipient
      properties:
        guardian:
          $ref: '#/components/schemas/Guardian'
        customerCenter:
          type: object
          additionalProperties: false
          required:
          - name
          - phoneNumber
          properties:
            name:
              type: string
            phoneNumber:
              $ref: '#/components/schemas/PhoneNumber'
        notificationRecipient:
          type: string
          enum:
          - SELF
...
    ErrorCode:
      type: string
      enum:
      - INVALID_REQUEST
      - INVALID_DATE_RANGE
      - INVALID_OAUTH_STATE
      - SESSION_REQUIRED
      - CSRF_INVALID
      - USAGE_CONSENT_REQUIRED
      - GUARDIAN_SHARE_CONSENT_REQUIRED
      - NOT_FOUND
      - SOURCE_ACCOUNT_NOT_FOUND
      - RECIPIENT_ACCOUNT_NOT_FOUND
      - REGISTERED_PERSON_NOT_FOUND
      - TRANSFER_RESULT_NOT_FOUND
      - STALE_REVISION
      - REQUEST_ID_CONFLICT
      - ACCOUNT_ALREADY_EXISTS
      - PATTERN_NUMBER_CONFLICT
      - PATTERN_LIMIT_EXCEEDED
      - PATTERN_INACTIVE
      - CONSENT_NOT_COMPLETED
      - ACCOUNT_NOT_READY
      - EXECUTION_ENDED
      - VISIT_ENDED
      - TRANSFER_PENDING
      - ANOMALY_NOT_ELIGIBLE
      - RECORDING_TEXT_CHANGED
      - SUGGESTION_CHANGED
      - SUGGESTION_IDENTICAL
      - AUDIO_TOO_LARGE
      - UNSUPPORTED_AUDIO_TYPE
      - PIN_MISMATCH
      - INSUFFICIENT_BALANCE
      - AMOUNT_TEXT_MISMATCH
      - INVALID_RECIPIENT
      - INVALID_TEMPLATE
      - INVALID_EVENT_TIME
      - INVALID_ACTION_COUNTS
      - AUDIO_DURATION_EXCEEDED
      - EMPTY_AUDIO
      - SWAP_CONFIRMATION_REQUIRED
      - INTERNAL_ERROR
      - KAKAO_AUTH_FAILED
      - TTS_PROVIDER_FAILED
      - NOTIFICATION_RESULT_UNKNOWN
      - DATABASE_UNAVAILABLE
      - DATABASE_NOT_CONFIGURED
      - KAKAO_NOT_CONFIGURED
      - DEMO_USER_CAPACITY_EXCEEDED
      - TTS_NOT_CONFIGURED
      - FILE_STORE_UNAVAILABLE
      - SUPPORT_NOT_CONFIGURED
  responses:
    Error400:
      description: 입력 형식·필수 필드 오류
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error401:
      description: 세션 없음 또는 만료
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error403:
      description: CSRF 또는 동의/업무 준비 조건 불충족
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error404:
      description: 대상 없음 또는 다른 사용자 소유. 존재 여부를 구분해 노출하지 않습니다.
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error409:
      description: 동시 수정·중복 키의 다른 요청·상태 충돌
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Error422:
      description: 형식은 맞지만 업무 검증 실패
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error500:
      description: 서버 처리 실패. 성공을 추정하거나 자동 재송금하지 않습니다.
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error502:
      description: 외부 제공사 실패
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Error503:
      description: 실행 설정·DB·외부 연동 준비 안 됨
      headers:
        Cache-Control:
          schema:
            type: string
            enum:
            - no-store
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...